Core:  - Added read-only fields support in reflection.  - Improved the JSON serializer to support all the data types.  - Added a visitor to help walk through the reflection information.  - Added a couple more path manipulation helper functions.  - Fixed hierarchy testing in RTTI tests.
Resources:
 - Simplified the resource loaders interface.

Pipeline:
 - Added resource processors query support.

Engine:
 - First pass at removing reliance on shared_ptrs.

Editor:
 - Property grids now work with all reflected objects.
 - Resources can now be selected and inspected.
 - Hooked up the new loading and saving support.
 - Fixed misc. stuff.

+++ inc/vapor/resources/Sound.h
@@ -30,10 +30,10 @@ public:
 	ACESSOR(Frequency, const int, frequency)
 
 	// Gets/sets the audio channels.
-	ACESSOR(Channels, byte, channels)
+	ACESSOR(Channels, int, channels)
 
 	// Gets/sets the audio sample size.
-	ACESSOR(Size, byte, size)
+	ACESSOR(Size, int, size)
 
 	// Gets/sets the buffer containing the data.
 	ACESSOR(Buffer, const std::vector<byte>&, dataBuffer)
@@ -47,10 +47,10 @@ protected:
 	int frequency;
 
 	// Sound channels.
-	byte channels;
+	int channels;
 
 	// Sound sample size.
-	byte size;
+	int size;
 
 	// Holds the sound data.
 	std::vector<byte> dataBuffer;
